+++ src/nvim/vterm/README.md
@@ -1,0 +1,1 @@
+Adopted from [libvterm](https://www.leonerd.org.uk/code/libvterm/)
